feat (contracts/DOS) make the codebase compatible with slither
* remove used defined types
* remove top-level `use * for *` and move them to contracts

Note on use defined types: the support for them has been added in 0.8.x, but for some reason slither complaining about them to not been defined. Maybe it's possible to support them with a better slither setup.
Note on top-leve `use * for *` there is an active PR on supporting this feature, so it's quite possible that we can add this feature back with the next slither release